Enquiry submit and review added.
Enquiries are submitted on the contact page and then the submitted enquiry
is displayed to the user (or an error message if the enquiry failed to add
to the database).
Next to add - data validation in the enquiry model.
Still to add - email to user confirming receipt of enquiry and email
notification that a new enquiry has been submitted.
Recaptcha check on contact form to be added too.